COMP Explicit casting to int in LevelWindow
* since GetBpe returns a size_t (unsigned long) the windows compiler
complains about ambiguous pow() call

+++ Core/Code/DataManagement/mitkLevelWindow.cpp
@@ -288,7 +288,8 @@ void mitk::LevelWindow::SetAuto(const mitk::Image* image, bool tryPicTags, bool 
   // Fix for bug# 344 Level Window wird bei Eris Cut bildern nicht richtig gesetzt
   if (image->GetPixelType()== typeid(int)  && image->GetPixelType().GetBpe() >= 8)
   {
-    if (minValue == -(pow((double)2.0, image->GetPixelType().GetBpe())/2))
+    // the windows compiler complains about ambiguos 'pow' call, therefore explicit casting to (double, int)
+    if (minValue == -(pow((double)2.0, (int) image->GetPixelType().GetBpe())/2))
     {
       minValue = min2ndValue;
     }
